// src/app/shared/models/model.ts
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface instructorAnnouncement {
  sNo?:number;
  date: string;
  course: string;
  announcement: string;
  status: string;
}